extract from utils
